requirements.txt
discord.py
gspread
google-auth
dotenv